woohoo

+++ actions/move-contents/main.py
@@ -4,6 +4,6 @@ SRC_PATH = sys.argv[1]
 DEST_PATH = sys.argv[2]
 
 if __name__ == '__main__':
-    shutil.rmtree(DEST_PATH)
-    os.makedirs(DEST_PATH)
+    if (os.path.isdir(DEST_PATH)):
+        shutil.rmtree(DEST_PATH)
     shutil.move(SRC_PATH, DEST_PATH, copy_function=shutil.copy_tree)
